[refactor] 헤더 리팩토링 3차
