+++ src/styles/reset.css
@@ -130,4 +130,7 @@ button {
   -webkit-appearance: none;
   -moz-appearance: none;
   appearance: none;
+  border: 0;
+  background-color: transparent;
+  cursor: pointer;
 }
